modify axios interceptor for ltsa
